functions/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-generativeai==0.3.2
flask==2.3.3
